+++ pyproject.toml
@@ -22,13 +22,13 @@ packages = ["src/goose"]
 goose-ai = "goose.module_name"
 
 [project.entry-points."goose.toolkit"]
-developer = "goose.toolkit.developer:Developer"
-github = "goose.toolkit.github:Github"
-screen = "goose.toolkit.screen:Screen"
-repo_context = "goose.toolkit.repo_context.repo_context:RepoContext"
+developer = "goose._internal.toolkit.developer:Developer"
+github = "goose._internal.toolkit.github:Github"
+screen = "goose._internal.toolkit.screen:Screen"
+repo_context = "goose._internal.toolkit.repo_context.repo_context:RepoContext"
 
 [project.entry-points."goose.profile"]
-default = "goose._internal.profile_config:default_profile"
+default = "goose._internal.profile:default_profile"
 
 [project.entry-points."goose.command"]
 file = "goose.command.file:FileCommand"
