+++ tests/_internal/profile/test_config.py
@@ -1,19 +1,20 @@
 from unittest.mock import patch
 
 import pytest
-from goose._internal.profile_config import _ensure_config, _read_config, _write_config, default_profile
+from goose._internal.profile import default_profile
+from goose._internal.profile.config import _ensure_config, _read_config, _write_config
 
 
 @pytest.fixture
 def mock_profile_config_path(tmp_path):
-    with patch("goose._internal.profile_config.PROFILES_CONFIG_PATH", tmp_path / "profiles.yaml") as mock_path:
+    with patch("goose._internal.profile.config.PROFILES_CONFIG_PATH", tmp_path / "profiles.yaml") as mock_path:
         yield mock_path
 
 
 @pytest.fixture
 def mock_default_model_configuration():
     with patch(
-        "goose._internal.profile_config._default_model_configuration", return_value=("provider", "processor", "accelerator")
+        "goose._internal.profile.config._default_model_configuration", return_value=("provider", "processor", "accelerator")
     ) as mock_default_model_configuration:
         yield mock_default_model_configuration
 
@@ -26,7 +27,6 @@ def test_read_write_config(mock_profile_config_path, profile_factory):
 
     assert _read_config() == profiles
 
-
 def test_ensure_config_create_profiles_file_with_default_profile(
     mock_profile_config_path, mock_default_model_configuration
 ):
@@ -37,7 +37,7 @@ def test_ensure_config_create_profiles_file_with_default_profile(
 
     assert _read_config() == {"default": default_profile(*mock_default_model_configuration())}
 
-
+@patch("goose._internal.profile.config.print")
 def test_ensure_config_add_default_profile(mock_profile_config_path, profile_factory, mock_default_model_configuration):
     existing_profile = profile_factory({"provider": "providerA"})
     _write_config({"profile1": existing_profile})
@@ -50,8 +50,8 @@ def test_ensure_config_add_default_profile(mock_profile_config_path, profile_fac
     }
 
 
-@patch("goose._internal.profile_config.Confirm.ask", return_value=True)
-@patch("goose._internal.profile_config.print")
+@patch("goose._internal.profile.config.Confirm.ask", return_value=True)
+@patch("goose._internal.profile.config.print")
 def test_ensure_config_overwrite_default_profile(
     mock_confirm, mock_profile_config_path, profile_factory, mock_default_model_configuration
 ):
@@ -64,8 +64,8 @@ def test_ensure_config_overwrite_default_profile(
     assert _read_config() == {"default": expected_default_profile}
 
 
-@patch("goose._internal.profile_config.Confirm.ask", return_value=False)
-@patch("goose._internal.profile_config.print")
+@patch("goose._internal.profile.config.Confirm.ask", return_value=False)
+@patch("goose._internal.profile.config.print")
 def test_ensure_config_keep_original_default_profile(
     mock_confirm, mock_profile_config_path, profile_factory, mock_default_model_configuration
 ):
